+++ CHANGELOG.md
@@ -3,12 +3,13 @@
 ## UNRELEASED
 
 * API:
-  * Updated `validatePhoneNumber` and `parsePhoneNumber` to make `regionCode` parameter optional
+  * Updated [`validatePhoneNumber`](./README.md#validatephonenumber) and [`parsePhoneNumber`](./README.md#parsephonenumber) to make `regionCode` parameter optional
     * `validatePhoneNumber`: If `regionCode` is omitted, the phone number is validated based on the region inferred from the phone number itself
     * `parsePhoneNumber`: The `regionCode` can be safely omitted if the phone number string contains "+" followed by the country code (e.g., international, E.164, or RFC3966 formats)
     * Note that the handler needs to have already been instantiated with metadata for the expected region(s)
-  * Added `inferPhoneNumberType` and `inferPhoneNumberRegion` methods
-  * Added `inputDigitAndRememberPosition` and `getRememberedPosition` methods to the AsYouTypeFormatter
+  * Added [`inferPhoneNumberType`](./README.md#inferphonenumbertype) and [`inferPhoneNumberRegion`](./README.md#inferphonenumberregion) methods
+  * Added [`validateLength`](./README.md#validatelength) method to perform length-only validations
+  * Added [`inputDigitAndRememberPosition`](./README.md#inputdigitandrememberposition) and [`getRememberedPosition`](./README.md#getrememberedposition) methods to the AsYouTypeFormatter
 
 ## 1.88.0